Seems that i have put the code to follow the CakePHP standard
To be honest i do not remember of these changes they were here at my local checkout,
and i do not know what exactly i have done in this code, but looks that i have put into the cakephp standard.

So i gonna push this as a develop branch of plugin, that's it! ;)